Based on RobR, Sam and RobL's suggestions, add a flag to the sys_mntent structure to enable/disable checking of the config files integrity. Right now, client-core sets this to 0, effectively skipping the fs.conf integrity checks and the pvfs2- system interface users (namely ping, etc) set this to 1 by default to enable the fs.conf integrity checks... Also fixed a few warnings on x86
SVN revision 5646

+++ src/proto/PINT-le-bytefield.c
@@ -579,7 +579,7 @@ static int lebf_decode_req(
     if (ptr != (char *) input_buffer + input_size)
     {
 	gossip_lerr("%s: op %d consumed %ld bytes, but message was %d bytes.\n",
-                    __func__, req->op, ptr - (char *) input_buffer, input_size);
+                    __func__, req->op, (long)(ptr - (char *) input_buffer), input_size);
 	ret = -PVFS_EPROTO;
     }
 
@@ -667,7 +667,7 @@ static int lebf_decode_resp(
 
     if (ptr != (char *) input_buffer + input_size) {
 	gossip_lerr("%s: op %d consumed %ld bytes, but message was %d bytes.\n",
-                    __func__, resp->op, ptr - (char *) input_buffer,
+                    __func__, resp->op, (long)(ptr - (char *) input_buffer),
                     input_size);
 	ret = -PVFS_EPROTO;
     }
